+++ src/main/java/org/prgrms/springbootbasic/service/VoucherService.java
@@ -1,7 +1,7 @@
 package org.prgrms.springbootbasic.service;
 
-import org.prgrms.springbootbasic.Request;
-import org.prgrms.springbootbasic.Response;
+import org.prgrms.springbootbasic.message.Request;
+import org.prgrms.springbootbasic.message.Response;
 import org.prgrms.springbootbasic.type.TypeOption;
 import org.prgrms.springbootbasic.factory.VoucherFactory;
 import org.prgrms.springbootbasic.repository.VoucherRepository;
